fixed loop bug
