+++ client/src/components/AudioPlayer.js
@@ -1,11 +1,11 @@
 import React, { useEffect, useRef, useState } from 'react';
-import { useAudioPlayer } from './AudioPlayerContext'; // Import the context
+import { useAudioPlayer } from './AudioPlayerContext';
 
 const AudioPlayer = () => {
-    const { ayahUrl, isPlaying, setIsPlaying, currentAyah, loopCount, playCount, setPlayCount } = useAudioPlayer();
+    const { ayahUrl, isPlaying, setIsPlaying, currentAyah } = useAudioPlayer();
     const audioRef = useRef(null);
-    const [currentTime, setCurrentTime] = useState(0);
-    const [duration, setDuration] = useState(0);
+    const [ currentTime, setCurrentTime ] = useState(0);
+    const [ duration, setDuration ] = useState(0);
 
     useEffect(() => {
         const audioElement = audioRef.current;
@@ -20,12 +20,12 @@ const AudioPlayer = () => {
             };
 
             const handleEnded = () => {
-                if (playCount < loopCount - 1) {
-                    setPlayCount((prevCount) => prevCount + 1);
+                console.log(`Audio ended for Ayah: ${currentAyah}`);
+                setIsPlaying(2);
+
+                if (audioElement && isPlaying === 1) {
+                    audioElement.currentTime = 0;
                     audioElement.play();
-                } else {
-                    setIsPlaying(false);
-                    setPlayCount(0); // Reset play count when done looping
                 }
             };
 
@@ -39,26 +39,29 @@ const AudioPlayer = () => {
                 audioElement.removeEventListener('ended', handleEnded);
             };
         }
-    }, [ayahUrl, isPlaying, setIsPlaying, playCount, loopCount]);
+    }, [ayahUrl, isPlaying, setIsPlaying, currentAyah]);
 
     useEffect(() => {
         const audioElement = audioRef.current;
+        if (isPlaying === 0) {
+            audioElement.pause();
+        }
 
         if (audioElement && ayahUrl) {
             if (currentAyah) {
                 if (audioElement.src !== ayahUrl) {
                     audioElement.src = ayahUrl;
                     audioElement.play();
+                    setIsPlaying(1);
                 } else {
-                    if (isPlaying) {
+                    if (isPlaying === 1) {
                         audioElement.play();
-                    } else {
-                        audioElement.pause();
                     }
                 }
             } else {
                 audioElement.src = ayahUrl;
                 audioElement.play();
+                setIsPlaying(1);
             }
         }
     }, [ayahUrl, isPlaying, currentAyah]);
@@ -68,10 +71,10 @@ const AudioPlayer = () => {
 
         if (audioElement.paused) {
             audioElement.play();
-            setIsPlaying(true);
+            setIsPlaying(1);
         } else {
             audioElement.pause();
-            setIsPlaying(false);
+            setIsPlaying(0);
         }
     };
 
@@ -88,39 +91,39 @@ const AudioPlayer = () => {
     };
 
     return (
-        <div className="single-player-2">
-            <input
-                type="range"
-                id="seekBar"
-                value={(currentTime / duration) * 100 || 0}
-                onChange={handleSeek}
-            />
-            <div className="timeRange">
-                <div className="currentTime" id="currentTime">
-                    {formatTime(currentTime)}
-                </div>
-                <button
-                    id="playPauseBtn"
-                    className="playPauseBtn icon-Btn"
-                    onClick={handlePlayPause}
-                    disabled={!ayahUrl} // Disable button if no ayahUrl
-                >
-                    <i className={`fa fa-${isPlaying ? 'pause' : 'play'} ${isPlaying ? 'pause-icon' : 'play-icon'}`}></i>
-                </button>
-                <div className="totalDuration" id="totalDuration">
-                    {formatTime(duration)}
-                </div>
-            </div>
-            <audio
-                id="audioPlayer"
-                ref={audioRef}
-                src=""
-                controls
-            >
-                <source src type="audio/mpeg" />
-                Your browser does not support the audio element.
-            </audio>
-        </div>
+      <div className="single-player-2">
+          <input
+            type="range"
+            id="seekBar"
+            value={(currentTime / duration) * 100 || 0}
+            onChange={handleSeek}
+          />
+          <div className="timeRange">
+              <div className="currentTime" id="currentTime">
+                  {formatTime(currentTime)}
+              </div>
+              <button
+                id="playPauseBtn"
+                className="playPauseBtn icon-Btn"
+                onClick={handlePlayPause}
+                disabled={!ayahUrl} // Disable button if no ayahUrl
+              >
+                  <i className={`fa fa-${isPlaying === 1 ? 'pause' : 'play'} ${isPlaying === 1 ? 'pause-icon' : 'play-icon'}`}></i>
+              </button>
+              <div className="totalDuration" id="totalDuration">
+                  {formatTime(duration)}
+              </div>
+          </div>
+          <audio
+            id="audioPlayer"
+            ref={audioRef}
+            src=""
+            controls
+          >
+              <source src type="audio/mpeg" />
+              Your browser does not support the audio element.
+          </audio>
+      </div>
     );
 };
