+++ client/src/components/AudioPlayerContext.js
@@ -6,12 +6,10 @@ export const useAudioPlayer = () => useContext(AudioPlayerContext);
 
 export const AudioPlayerProvider = ({ children }) => {
     const [ayahUrl, setAyahUrl] = useState('');
-    const [isPlaying, setIsPlaying] = useState(false);
+    const [isPlaying, setIsPlaying] = useState(0);
     const [currentAyah, setCurrentAyah] = useState(null);
-    const [loopCount, setLoopCount] = useState(10); // Add state for loop count
-    const [playCount, setPlayCount] = useState(0); // Add state for play count
 
-    return (
+  return (
         <AudioPlayerContext.Provider
             value={{
                 ayahUrl,
@@ -20,10 +18,6 @@ export const AudioPlayerProvider = ({ children }) => {
                 setIsPlaying,
                 currentAyah,
                 setCurrentAyah,
-                loopCount,
-                setLoopCount,
-                playCount,
-                setPlayCount // Ensure playCount state is provided
             }}
         >
             {children}
